Revert "test(eve): split Slack channel coverage"
This reverts commit e5b516c061ed4f656ebee9980aaeb7348ad66834.
