+++ packages/eve/src/public/channels/slack/slackChannel.test.ts
@@ -1,12 +1,29 @@
-import { afterAll, beforeEach, describe, expect, it, vi } from "vitest";
+import { createHmac } from "node:crypto";
 
+import { afterAll, afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+
+import { buildAdapterContext } from "#channel/adapter-context.js";
+import { callAdapterEventHandler, type ChannelAdapter } from "#channel/adapter.js";
 import { isCompiledChannel, type CompiledChannel } from "#channel/compiled-channel.js";
+import { isHttpRouteDefinition } from "#channel/routes.js";
+import { ContextContainer, contextStorage } from "#context/container.js";
+import { SessionKey } from "#context/keys.js";
+import type { HandleMessageStreamEvent } from "#protocol/message.js";
+import { decodeSlackApiBody } from "#public/channels/slack/api-encoding.js";
 import {
-  buildSignedSlackRequest as buildSignedRequest,
-  fireSlackPost as firePost,
-  parseSlackRequestBody,
-  SLACK_TEST_SIGNING_SECRET as SIGNING_SECRET,
-} from "#internal/testing/slack-channel-test-helpers.js";
+  buildHitlResponderBlockId,
+  HITL_ACTION_PREFIX,
+  HITL_FREEFORM_ACTION_PREFIX,
+  HITL_FREEFORM_MODAL_ACTION_ID,
+  HITL_FREEFORM_MODAL_BLOCK_ID,
+  HITL_FREEFORM_MODAL_CALLBACK_ID,
+  parseHitlResponderBinding,
+} from "#public/channels/slack/hitl.js";
+import {
+  SLACK_MESSAGE_TEXT_MAX_LENGTH,
+  SLACK_SECTION_TEXT_MAX_LENGTH,
+} from "#public/channels/slack/limits.js";
+import { defaultSlackAuth } from "#public/channels/slack/index.js";
 import {
   constrainAuthorizationRequired,
   slackChannel,
@@ -16,11 +33,224 @@ import {
 } from "#public/channels/slack/slackChannel.js";
 import type { SessionContext } from "#public/definitions/callback-context.js";
 
+function getAdapter(channel: unknown): ChannelAdapter<any> {
+  if (!isCompiledChannel(channel)) {
+    throw new Error("Expected a CompiledChannel.");
+  }
+  return channel.adapter;
+}
+
 function asCompiled<T = unknown>(channel: unknown): CompiledChannel<T> {
   if (!isCompiledChannel(channel)) {
     throw new Error("Expected a CompiledChannel.");
   }
   return channel as CompiledChannel<T>;
+}
+
+// Decodes a captured Slack outbound request body. All Slack API calls
+// are sent form-encoded, so the test infra mirrors that to inspect them.
+function parseSlackRequestBody(init: RequestInit | undefined): Record<string, unknown> {
+  if (!init?.body) return {};
+  const contentType = init.headers ? new Headers(init.headers).get("content-type") : null;
+  return decodeSlackApiBody(init.body, contentType) as Record<string, unknown>;
+}
+
+function makeResponderBinding(responderUserId: string, requestId: string) {
+  const binding = parseHitlResponderBinding(
+    buildHitlResponderBlockId({ requestId, responderUserId }),
+  );
+  if (!binding) throw new Error("Expected a valid responder binding.");
+  return binding;
+}
+
+function withState(
+  adapter: ChannelAdapter<any>,
+  state: Record<string, unknown>,
+): ChannelAdapter<any> {
+  return { ...adapter, state: { ...adapter.state, ...state } };
+}
+
+function stubAccessor() {
+  return { get: () => undefined, set: () => {} } as any;
+}
+
+const stubAlsContext = (() => {
+  const ctx = new ContextContainer();
+  ctx.setVirtualContext(SessionKey, {
+    sessionId: "test-session",
+    auth: { current: null, initiator: null },
+    turn: { id: "test-turn", sequence: 0 },
+  });
+  return ctx;
+})();
+
+function callEvent(
+  adapter: ChannelAdapter,
+  event: HandleMessageStreamEvent,
+  ctx: any,
+): Promise<HandleMessageStreamEvent> {
+  return contextStorage.run(stubAlsContext, () => callAdapterEventHandler(adapter, event, ctx));
+}
+
+/**
+ * Accessor whose `set` writes are captured so tests can assert on
+ * `setContinuationToken` flowing through the SessionHandle. Returns
+ * undefined for unset keys (matching the real `ContextContainer`
+ * behavior), while seeding the current continuation token so
+ * SessionHandle can preserve the runtime namespace.
+ */
+function captureAccessor(initialContinuationToken: string): {
+  accessor: any;
+  writes: Array<[string, unknown]>;
+} {
+  const writes: Array<[string, unknown]> = [];
+  let continuationToken = initialContinuationToken;
+  return {
+    writes,
+    accessor: {
+      get: (key: { name: string }) =>
+        key.name === "eve.continuationToken" ? continuationToken : undefined,
+      set: (key: { name: string }, value: unknown | ((current: unknown) => unknown)) => {
+        const next =
+          typeof value === "function" ? (value as (current: unknown) => unknown)(undefined) : value;
+        if (key.name === "eve.continuationToken") {
+          continuationToken = String(next);
+        }
+        writes.push([key.name, next]);
+        return next;
+      },
+    },
+  };
+}
+
+function makeEvent<T extends HandleMessageStreamEvent["type"]>(
+  type: T,
+  data: unknown,
+): HandleMessageStreamEvent {
+  return { type, data } as HandleMessageStreamEvent;
+}
+
+const THREAD_STATE = {
+  channelId: "C01",
+  threadTs: "1700000000.000001",
+  teamId: "T01",
+  triggeringUserId: "U_REQUESTER",
+};
+
+const SIGNING_SECRET = "test-signing-secret";
+
+function buildSignedRequest(input: {
+  body: string;
+  contentType?: string;
+  headers?: Record<string, string>;
+  timestamp?: number;
+  signingSecret?: string;
+}): Request {
+  const timestamp = input.timestamp ?? Math.floor(Date.now() / 1000);
+  const secret = input.signingSecret ?? SIGNING_SECRET;
+  const signature = `v0=${createHmac("sha256", secret)
+    .update(`v0:${timestamp}:${input.body}`)
+    .digest("hex")}`;
+  return new Request("https://example.com/eve/v1/slack", {
+    method: "POST",
+    headers: {
+      "content-type": input.contentType ?? "application/json",
+      "x-slack-request-timestamp": String(timestamp),
+      "x-slack-signature": signature,
+      ...input.headers,
+    },
+    body: input.body,
+  });
+}
+
+function buildSignedInteractionRequest(payload: Record<string, unknown>): Request {
+  const body = new URLSearchParams({ payload: JSON.stringify(payload) }).toString();
+  return buildSignedRequest({
+    body,
+    contentType: "application/x-www-form-urlencoded",
+  });
+}
+
+function buildHitlBlockActionPayload(input: {
+  readonly actorUserId: string;
+  readonly kind: "button" | "freeform";
+  readonly responderUserId: string;
+}): Record<string, unknown> {
+  const requestId = input.kind === "button" ? "approval_abc123" : "call_abc123";
+  const action =
+    input.kind === "button"
+      ? {
+          action_id: `${HITL_ACTION_PREFIX}${requestId}:button:0`,
+          block_id: buildHitlResponderBlockId({
+            requestId,
+            responderUserId: input.responderUserId,
+          }),
+          text: { type: "plain_text", text: "Approve" },
+          value: "approve",
+        }
+      : {
+          action_id: `${HITL_FREEFORM_ACTION_PREFIX}${requestId}`,
+          block_id: buildHitlResponderBlockId({
+            requestId,
+            responderUserId: input.responderUserId,
+          }),
+          text: { type: "plain_text", text: "Type your answer" },
+          value: requestId,
+        };
+
+  return {
+    actions: [action],
+    channel: { id: "C01" },
+    message: {
+      blocks: [{ type: "section", text: { type: "mrkdwn", text: "Approve?" } }],
+      thread_ts: "1700000000.000001",
+      ts: "1700000000.000010",
+    },
+    team: { id: "T01" },
+    trigger_id: "trigger-1",
+    type: "block_actions",
+    user: {
+      id: input.actorUserId,
+      name: "test-user",
+      team_id: "T01",
+      username: "test-user",
+    },
+  };
+}
+
+function buildFreeformSubmissionPayload(input: {
+  readonly actorUserId: string;
+  readonly responderUserId: string;
+  readonly text: string;
+}): Record<string, unknown> {
+  return {
+    team: { id: "T01" },
+    type: "view_submission",
+    user: {
+      id: input.actorUserId,
+      name: "test-user",
+      team_id: "T01",
+      username: "test-user",
+    },
+    view: {
+      callback_id: HITL_FREEFORM_MODAL_CALLBACK_ID,
+      private_metadata: JSON.stringify({
+        channelId: "C01",
+        continuationToken: "C01:1700000000.000001",
+        messageTs: "1700000000.000010",
+        requestId: "call_abc123",
+        responderBinding: makeResponderBinding(input.responderUserId, "call_abc123"),
+        threadTs: "1700000000.000001",
+      }),
+      state: {
+        values: {
+          [HITL_FREEFORM_MODAL_BLOCK_ID]: {
+            [HITL_FREEFORM_MODAL_ACTION_ID]: { value: input.text },
+          },
+        },
+      },
+    },
+  };
 }
 
 let mentionCounter = 0;
@@ -81,6 +311,575 @@ function buildDirectMessageBody(overrides?: {
   });
   return { body, channel, ts };
 }
+
+async function firePost(
+  channel: unknown,
+  request: Request,
+): Promise<{
+  response: Response;
+  send: ReturnType<typeof vi.fn>;
+  waitUntil: ReturnType<typeof vi.fn>;
+}> {
+  const compiled = asCompiled(channel);
+  const post = compiled.routes.find((r) => r.method === "POST");
+  if (!post || !isHttpRouteDefinition(post)) {
+    throw new Error("Expected slack channel to define a POST route.");
+  }
+  const send = vi.fn().mockResolvedValue({ id: "s1", continuationToken: "ct" });
+  const waitUntil = vi.fn();
+
+  const response = await post.handler(request, {
+    send,
+    waitUntil,
+    getSession: vi.fn() as any,
+    params: {},
+    requestIp: null,
+  } as any);
+
+  let drained = 0;
+  while (drained < waitUntil.mock.calls.length) {
+    const pending = waitUntil.mock.calls.slice(drained).map(([task]) => task as Promise<unknown>);
+    drained = waitUntil.mock.calls.length;
+    await Promise.allSettled(pending);
+  }
+
+  return { response, send, waitUntil };
+}
+
+describe("slackChannel() default event handlers", () => {
+  let fetchMock: ReturnType<typeof vi.fn>;
+  beforeEach(() => {
+    fetchMock = vi.fn().mockImplementation(() =>
+      Promise.resolve(
+        new Response(JSON.stringify({ ok: true, ts: "1700000001.000001" }), {
+          headers: { "content-type": "application/json" },
+        }),
+      ),
+    );
+    vi.stubGlobal("fetch", fetchMock);
+  });
+  afterEach(() => {
+    vi.useRealTimers();
+  });
+
+  it("message.completed posts the agent message via Slack API", async () => {
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+
+    await callEvent(
+      adapter,
+      makeEvent("message.completed", {
+        finishReason: "stop",
+        message: "Hello from the agent",
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [url, init] = fetchMock.mock.calls[0]!;
+    expect(String(url)).toBe("https://slack.com/api/chat.postMessage");
+    const body = parseSlackRequestBody(init as RequestInit);
+    expect(body).toMatchObject({
+      channel: "C01",
+      thread_ts: "1700000000.000001",
+      markdown_text: "Hello from the agent",
+    });
+  });
+
+  it("message.completed skips post when finishReason is tool-calls", async () => {
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+
+    await callEvent(
+      adapter,
+      makeEvent("message.completed", {
+        finishReason: "tool-calls",
+        message: "Should not post",
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).not.toHaveBeenCalled();
+  });
+
+  it("input.requested posts an approval card with Slack-unique button action ids", async () => {
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+
+    await callEvent(
+      adapter,
+      makeEvent("input.requested", {
+        requests: [
+          {
+            action: {
+              callId: "call_abc123",
+              input: { operation: "deleteMany" },
+              kind: "tool-call",
+              toolName: "mongodb-mutate",
+            },
+            display: "confirmation",
+            options: [
+              { id: "approve", label: "Yes" },
+              { id: "deny", label: "No" },
+            ],
+            prompt: "Approve tool call: mongodb-mutate",
+            requestId: "approval_abc123",
+          },
+        ],
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [url, init] = fetchMock.mock.calls[0]!;
+    expect(String(url)).toBe("https://slack.com/api/chat.postMessage");
+    const body = parseSlackRequestBody(init as RequestInit) as {
+      blocks: Array<{ elements?: Array<{ action_id: string; value: string }> }>;
+      channel: string;
+      text: string;
+      thread_ts: string;
+    };
+    expect(body).toMatchObject({
+      channel: "C01",
+      text: "Approve tool call: mongodb-mutate",
+      thread_ts: "1700000000.000001",
+    });
+
+    const actions = body.blocks.find((block) => Array.isArray(block.elements));
+    expect(actions).toMatchObject({
+      block_id: buildHitlResponderBlockId({
+        requestId: "approval_abc123",
+        responderUserId: "U_REQUESTER",
+      }),
+    });
+    const actionIds = actions?.elements?.map((element) => element.action_id) ?? [];
+    expect(actionIds).toEqual([
+      `${HITL_ACTION_PREFIX}approval_abc123:button:0`,
+      `${HITL_ACTION_PREFIX}approval_abc123:button:1`,
+    ]);
+    expect(new Set(actionIds).size).toBe(actionIds.length);
+  });
+
+  it("input.requested caps section and fallback text so Slack does not reject the post", async () => {
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+    const longPrompt = "x".repeat(SLACK_SECTION_TEXT_MAX_LENGTH + 500);
+
+    await callEvent(
+      adapter,
+      makeEvent("input.requested", {
+        requests: [
+          {
+            action: {
+              callId: "call_long",
+              input: {},
+              kind: "tool-call",
+              toolName: "ask_question",
+            },
+            display: "text",
+            prompt: longPrompt,
+            requestId: "call_long",
+          },
+        ],
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [, init] = fetchMock.mock.calls[0]!;
+    const body = parseSlackRequestBody(init as RequestInit) as {
+      blocks: Array<{ type: string; text?: { text: string } }>;
+      text: string;
+    };
+    const promptSection = body.blocks.find((block) => block.type === "section");
+    expect(promptSection?.text?.text.length).toBeLessThanOrEqual(SLACK_SECTION_TEXT_MAX_LENGTH);
+    expect(promptSection?.text?.text.endsWith("...")).toBe(true);
+    expect(body.text.length).toBeLessThanOrEqual(SLACK_MESSAGE_TEXT_MAX_LENGTH);
+  });
+
+  it("turn.started calls assistant.threads.setStatus", async () => {
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+
+    await callEvent(
+      adapter,
+      makeEvent("turn.started", { sequence: 0, stepIndex: 0, turnId: "t1" }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [url, init] = fetchMock.mock.calls[0]!;
+    expect(String(url)).toBe("https://slack.com/api/assistant.threads.setStatus");
+    const body = parseSlackRequestBody(init as RequestInit);
+    expect(body).toMatchObject({
+      channel_id: "C01",
+      thread_ts: "1700000000.000001",
+      status: "Working...",
+      loading_messages: ["Working..."],
+    });
+  });
+
+  it("reasoning.appended calls assistant.threads.setStatus with a truncated snippet", async () => {
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+    const longReasoning =
+      "Need to inspect the implementation and verify the Slack typing status behavior before editing.";
+
+    await callEvent(
+      adapter,
+      makeEvent("reasoning.appended", {
+        reasoningDelta: longReasoning,
+        reasoningSoFar: `${longReasoning}\nThen continue.`,
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [url, init] = fetchMock.mock.calls[0]!;
+    expect(String(url)).toBe("https://slack.com/api/assistant.threads.setStatus");
+    const body = parseSlackRequestBody(init as RequestInit);
+    expect(body).toMatchObject({
+      channel_id: "C01",
+      thread_ts: "1700000000.000001",
+    });
+    expect((body.status as string).length).toBeLessThanOrEqual(50);
+    expect((body.status as string).endsWith("...")).toBe(true);
+    expect(body.loading_messages).toEqual([body.status]);
+    expect(ctx.state.lastReasoningTypingStatus).toBe(body.status);
+  });
+
+  it("reasoning.appended immediately publishes progressive extensions of four characters", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date("2026-06-18T12:00:00Z"));
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+    const reasoningEvent = (reasoningDelta: string, reasoningSoFar: string) =>
+      makeEvent("reasoning.appended", {
+        reasoningDelta,
+        reasoningSoFar,
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      });
+
+    await callEvent(adapter, reasoningEvent("I", "I"), ctx);
+    await callEvent(adapter, reasoningEvent(" ca", "I ca"), ctx);
+    await callEvent(adapter, reasoningEvent("n", "I can"), ctx);
+
+    const statuses = fetchMock.mock.calls.map(
+      ([, init]) => parseSlackRequestBody(init as RequestInit).status,
+    );
+    expect(statuses).toEqual(["I", "I can"]);
+  });
+
+  it("reasoning.appended requires a matching prefix and four new characters", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date("2026-06-18T12:00:00Z"));
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+    const reasoningEvent = (reasoningSoFar: string) =>
+      makeEvent("reasoning.appended", {
+        reasoningDelta: reasoningSoFar,
+        reasoningSoFar,
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      });
+
+    await callEvent(adapter, reasoningEvent("Need"), ctx);
+    vi.setSystemTime(new Date("2026-06-18T12:00:01Z"));
+    await callEvent(adapter, reasoningEvent("Need to"), ctx);
+    await callEvent(adapter, reasoningEvent("Check something else"), ctx);
+    vi.setSystemTime(new Date("2026-06-18T12:00:05Z"));
+    await callEvent(adapter, reasoningEvent("Need to"), ctx);
+
+    expect(fetchMock).toHaveBeenCalledTimes(2);
+    const statuses = fetchMock.mock.calls.map(
+      ([, init]) => parseSlackRequestBody(init as RequestInit).status,
+    );
+    expect(statuses).toEqual(["Need", "Need to"]);
+  });
+
+  it("turn.started resets reasoning status throttling", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date("2026-06-18T12:00:00Z"));
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+
+    await callEvent(
+      adapter,
+      makeEvent("reasoning.appended", {
+        reasoningDelta: "Need to inspect the repo.",
+        reasoningSoFar: "Need to inspect the repo.",
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      }),
+      ctx,
+    );
+    vi.setSystemTime(new Date("2026-06-18T12:00:01Z"));
+    await callEvent(
+      adapter,
+      makeEvent("turn.started", { sequence: 0, stepIndex: 0, turnId: "t2" }),
+      ctx,
+    );
+    await callEvent(
+      adapter,
+      makeEvent("reasoning.appended", {
+        reasoningDelta: "Fresh turn reasoning.",
+        reasoningSoFar: "Fresh turn reasoning.",
+        sequence: 1,
+        stepIndex: 0,
+        turnId: "t2",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(3);
+    const statuses = fetchMock.mock.calls.map(
+      ([, init]) => parseSlackRequestBody(init as RequestInit).status,
+    );
+    expect(statuses).toEqual(["Need to inspect the repo.", "Working...", "Fresh turn reasoning."]);
+  });
+
+  it("session.failed posts a terminal markdown message with error hint and id", async () => {
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+
+    await callEvent(
+      adapter,
+      makeEvent("session.failed", {
+        code: "internal",
+        details: { errorId: "abc-123", name: "WorkflowExecutionFailed" },
+        message: "boom",
+        sessionId: "s1",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const body = parseSlackRequestBody(fetchMock.mock.calls[0]![1] as RequestInit);
+    expect(body.markdown_text).toContain("couldn't recover");
+    expect(body.markdown_text).toContain("Start a new thread");
+    expect(body.markdown_text).toContain("WorkflowExecutionFailed");
+    expect(body.markdown_text).toContain("abc-123");
+  });
+
+  it("actions.requested typing indicator is truncated to Slack's length cap", async () => {
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      THREAD_STATE,
+    );
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+
+    const longTool = "search_internal_documentation_for_relevant_passages";
+
+    await callEvent(
+      adapter,
+      makeEvent("actions.requested", {
+        actions: [
+          { kind: "tool-call", toolName: longTool, callId: "c1", input: {} },
+          { kind: "tool-call", toolName: longTool, callId: "c2", input: {} },
+          { kind: "tool-call", toolName: longTool, callId: "c3", input: {} },
+        ],
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const body = parseSlackRequestBody(fetchMock.mock.calls[0]![1] as RequestInit);
+    expect((body.status as string).length).toBeLessThanOrEqual(50);
+    expect((body.status as string).endsWith("...")).toBe(true);
+  });
+});
+
+describe("rebuildSlackContext", () => {
+  beforeEach(() => {
+    vi.stubGlobal("fetch", vi.fn());
+  });
+
+  it("seeds ctx.slack with state channelId / threadTs / teamId", () => {
+    const adapter = withState(getAdapter(slackChannel()), THREAD_STATE);
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+    expect(ctx.slack.channelId).toBe("C01");
+    expect(ctx.slack.threadTs).toBe("1700000000.000001");
+    expect(ctx.slack.teamId).toBe("T01");
+    expect(ctx.session).toBeDefined();
+    expect("threadId" in ctx.thread).toBe(false);
+    expect("id" in ctx.thread).toBe(false);
+  });
+
+  it("falls back to empty strings when state has no thread", () => {
+    const adapter = withState(getAdapter(slackChannel()), {
+      channelId: null,
+      threadTs: null,
+      teamId: null,
+    });
+    const ctx = buildAdapterContext(adapter, stubAccessor());
+    expect(ctx.slack.channelId).toBe("");
+    expect(ctx.slack.threadTs).toBe("");
+    expect("threadId" in ctx.thread).toBe(false);
+  });
+
+  it("auto-anchors state.threadTs and re-keys the session on the first post", async () => {
+    const fetchMock = vi.fn().mockResolvedValue(
+      new Response(JSON.stringify({ ok: true, ts: "1800000000.123456" }), {
+        headers: { "content-type": "application/json" },
+      }),
+    );
+    vi.stubGlobal("fetch", fetchMock);
+
+    const adapter = withState(
+      getAdapter(slackChannel({ credentials: { botToken: "xoxb-test" } })),
+      {
+        channelId: "C01",
+        threadTs: null,
+        teamId: null,
+      },
+    );
+    const { accessor, writes } = captureAccessor("slack:C01:");
+    const ctx = buildAdapterContext(adapter, accessor);
+
+    // Fire message.completed → default handler posts the agent message.
+    await callEvent(
+      adapter,
+      makeEvent("message.completed", {
+        finishReason: "stop",
+        message: "Daily digest",
+        sequence: 0,
+        stepIndex: 0,
+        turnId: "t1",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const firstBody = parseSlackRequestBody(fetchMock.mock.calls[0]![1] as RequestInit);
+    expect(firstBody.thread_ts).toBeUndefined();
+    expect((adapter.state as { threadTs: string | null }).threadTs).toBe("1800000000.123456");
+
+    // The anchor moment wrote the new continuation token to context
+    // via `session.setContinuationToken(...)`. The workflow body picks
+    // this up via `reconcileSessionContinuationToken` after the step.
+    const tokenWrites = writes.filter(([key]) => key === "eve.continuationToken");
+    expect(tokenWrites).toEqual([["eve.continuationToken", "slack:C01:1800000000.123456"]]);
+
+    // A follow-up message.completed now threads under the anchor.
+    fetchMock.mockResolvedValueOnce(
+      new Response(JSON.stringify({ ok: true, ts: "1800000000.999999" }), {
+        headers: { "content-type": "application/json" },
+      }),
+    );
+    await callEvent(
+      adapter,
+      makeEvent("message.completed", {
+        finishReason: "stop",
+        message: "Follow-up detail",
+        sequence: 1,
+        stepIndex: 1,
+        turnId: "t2",
+      }),
+      ctx,
+    );
+
+    expect(fetchMock).toHaveBeenCalledTimes(2);
+    const secondBody = parseSlackRequestBody(fetchMock.mock.calls[1]![1] as RequestInit);
+    expect(secondBody.thread_ts).toBe("1800000000.123456");
+
+    // Once anchored, setContinuationToken does not fire again — the
+    // raw token is unchanged across subsequent posts.
+    const allTokenWrites = writes.filter(([key]) => key === "eve.continuationToken");
+    expect(allTokenWrites).toHaveLength(1);
+  });
+});
+
+describe("defaultSlackAuth", () => {
+  it("is exported from the public Slack entry point", () => {
+    const auth = defaultSlackAuth(
+      {
+        attachments: [],
+        author: {
+          fullName: undefined,
+          isBot: false,
+          isMe: false,
+          userId: "U01",
+          userName: "ada",
+        },
+        channelId: "C01",
+        markdown: "hello",
+        raw: {},
+        teamId: "T01",
+        text: "hello",
+        threadTs: "1700000000.000001",
+        ts: "1700000000.000002",
+      },
+      {
+        slack: { channelId: "C01", threadTs: "1700000000.000001", teamId: "T01" } as never,
+        thread: {} as never,
+      },
+    );
+
+    expect(auth).toMatchObject({
+      attributes: {
+        channel_id: "C01",
+        thread_ts: "1700000000.000001",
+        user_id: "U01",
+        user_name: "ada",
+      },
+      authenticator: "slack-webhook",
+      principalId: "slack:T01:U01",
+      principalType: "user",
+    });
+  });
+});
 
 describe("slackChannel() inbound mention pipeline", () => {
   const ORIGINAL_SIGNING_SECRET = process.env.SLACK_SIGNING_SECRET;
@@ -483,6 +1282,288 @@ describe("slackChannel() inbound direct message pipeline", () => {
 
     expect(onDirectMessage).toHaveBeenCalledTimes(1);
     expect(send).not.toHaveBeenCalled();
+  });
+});
+
+describe("slackChannel() HITL interaction pipeline", () => {
+  const ORIGINAL_SIGNING_SECRET = process.env.SLACK_SIGNING_SECRET;
+  const ORIGINAL_BOT_TOKEN = process.env.SLACK_BOT_TOKEN;
+  let fetchMock: ReturnType<typeof vi.fn>;
+
+  beforeEach(() => {
+    process.env.SLACK_SIGNING_SECRET = SIGNING_SECRET;
+    fetchMock = vi.fn().mockResolvedValue(
+      new Response(JSON.stringify({ ok: true, ts: "1700000001.000001" }), {
+        headers: { "content-type": "application/json" },
+      }),
+    );
+    vi.stubGlobal("fetch", fetchMock);
+  });
+
+  afterAll(() => {
+    if (ORIGINAL_SIGNING_SECRET === undefined) {
+      delete process.env.SLACK_SIGNING_SECRET;
+    } else {
+      process.env.SLACK_SIGNING_SECRET = ORIGINAL_SIGNING_SECRET;
+    }
+    if (ORIGINAL_BOT_TOKEN === undefined) {
+      delete process.env.SLACK_BOT_TOKEN;
+    } else {
+      process.env.SLACK_BOT_TOKEN = ORIGINAL_BOT_TOKEN;
+    }
+  });
+
+  it("resumes HITL button answers with the approving Slack user auth", async () => {
+    const channel = slackChannel({ credentials: { botToken: "xoxb-test" } });
+
+    const { send } = await firePost(
+      channel,
+      buildSignedInteractionRequest({
+        type: "block_actions",
+        team: { id: "T01" },
+        user: {
+          id: "U_APPROVER",
+          username: "ada",
+          name: "ada",
+          team_id: "T01",
+        },
+        channel: { id: "C01" },
+        message: {
+          ts: "1700000000.000010",
+          thread_ts: "1700000000.000001",
+          blocks: [{ type: "section", text: { type: "mrkdwn", text: "Approve?" } }],
+        },
+        actions: [
+          {
+            action_id: `${HITL_ACTION_PREFIX}approval_abc123:button:0`,
+            block_id: buildHitlResponderBlockId({
+              requestId: "approval_abc123",
+              responderUserId: "U_APPROVER",
+            }),
+            text: { type: "plain_text", text: "Approve" },
+            value: "approve",
+          },
+        ],
+      }),
+    );
+
+    expect(send).toHaveBeenCalledTimes(1);
+    const [payload, options] = send.mock.calls[0]!;
+    expect(payload).toEqual({
+      inputResponses: [{ optionId: "approve", requestId: "approval_abc123" }],
+    });
+    expect(options).toMatchObject({
+      auth: {
+        attributes: {
+          author_type: "user",
+          channel_id: "C01",
+          team_id: "T01",
+          thread_ts: "1700000000.000001",
+          user_id: "U_APPROVER",
+          user_name: "ada",
+        },
+        authenticator: "slack-webhook",
+        issuer: "slack:T01",
+        principalId: "slack:T01:U_APPROVER",
+        principalType: "user",
+      },
+      continuationToken: "C01:1700000000.000001",
+      state: {
+        channelId: "C01",
+        teamId: "T01",
+        threadTs: "1700000000.000001",
+        triggeringUserId: "U_APPROVER",
+      },
+    });
+  });
+
+  it("rejects HITL button answers from a different Slack user", async () => {
+    const channel = slackChannel({ credentials: { botToken: "xoxb-test" } });
+
+    const { send } = await firePost(
+      channel,
+      buildSignedInteractionRequest(
+        buildHitlBlockActionPayload({
+          actorUserId: "U_INTRUDER",
+          kind: "button",
+          responderUserId: "U_OWNER",
+        }),
+      ),
+    );
+
+    expect(send).not.toHaveBeenCalled();
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [url, init] = fetchMock.mock.calls[0]!;
+    expect(String(url)).toBe("https://slack.com/api/chat.postEphemeral");
+    expect(parseSlackRequestBody(init as RequestInit)).toMatchObject({
+      channel: "C01",
+      thread_ts: "1700000000.000001",
+      user: "U_INTRUDER",
+    });
+  });
+
+  it("explains that an unbound HITL card must be recreated", async () => {
+    const channel = slackChannel({ credentials: { botToken: "xoxb-test" } });
+    const payload = buildHitlBlockActionPayload({
+      actorUserId: "U_OWNER",
+      kind: "button",
+      responderUserId: "U_OWNER",
+    });
+    const action = (payload.actions as Array<Record<string, unknown>>)[0];
+    if (!action) throw new Error("Expected an interaction action.");
+    action.block_id = "legacy-actions";
+
+    const { send } = await firePost(channel, buildSignedInteractionRequest(payload));
+
+    expect(send).not.toHaveBeenCalled();
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [url, init] = fetchMock.mock.calls[0]!;
+    expect(String(url)).toBe("https://slack.com/api/chat.postEphemeral");
+    expect(parseSlackRequestBody(init as RequestInit)).toMatchObject({
+      channel: "C01",
+      markdown_text:
+        "This prompt is no longer valid. Start a new request so it can be bound to your Slack identity.",
+      thread_ts: "1700000000.000001",
+      user: "U_OWNER",
+    });
+  });
+
+  it("does not open a freeform modal for a different Slack user", async () => {
+    const channel = slackChannel({ credentials: { botToken: "xoxb-test" } });
+
+    const { send } = await firePost(
+      channel,
+      buildSignedInteractionRequest(
+        buildHitlBlockActionPayload({
+          actorUserId: "U_INTRUDER",
+          kind: "freeform",
+          responderUserId: "U_OWNER",
+        }),
+      ),
+    );
+
+    expect(send).not.toHaveBeenCalled();
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    expect(String(fetchMock.mock.calls[0]![0])).toBe("https://slack.com/api/chat.postEphemeral");
+  });
+
+  it("opens a freeform modal for the bound Slack user", async () => {
+    const channel = slackChannel({ credentials: { botToken: "xoxb-test" } });
+
+    const { send } = await firePost(
+      channel,
+      buildSignedInteractionRequest(
+        buildHitlBlockActionPayload({
+          actorUserId: "U_OWNER",
+          kind: "freeform",
+          responderUserId: "U_OWNER",
+        }),
+      ),
+    );
+
+    expect(send).not.toHaveBeenCalled();
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [url, init] = fetchMock.mock.calls[0]!;
+    expect(String(url)).toBe("https://slack.com/api/views.open");
+    const requestBody = JSON.parse(String((init as RequestInit).body)) as {
+      view: { private_metadata: string };
+    };
+    const metadata = JSON.parse(requestBody.view.private_metadata) as Record<string, unknown>;
+    expect(metadata).toMatchObject({
+      requestId: "call_abc123",
+      responderBinding: makeResponderBinding("U_OWNER", "call_abc123"),
+    });
+    expect(metadata).not.toHaveProperty("responderUserId");
+  });
+
+  it("resumes freeform modal answers with the submitting Slack user auth", async () => {
+    const channel = slackChannel({ credentials: { botToken: "xoxb-test" } });
+
+    const { send } = await firePost(
+      channel,
+      buildSignedInteractionRequest({
+        type: "view_submission",
+        team: { id: "T01" },
+        user: {
+          id: "U_SUBMITTER",
+          username: "grace",
+          name: "grace",
+          team_id: "T01",
+        },
+        view: {
+          callback_id: HITL_FREEFORM_MODAL_CALLBACK_ID,
+          private_metadata: JSON.stringify({
+            channelId: "C01",
+            continuationToken: "C01:1700000000.000001",
+            messageTs: "1700000000.000010",
+            requestId: "call_abc123",
+            responderBinding: makeResponderBinding("U_SUBMITTER", "call_abc123"),
+            threadTs: "1700000000.000001",
+          }),
+          state: {
+            values: {
+              [HITL_FREEFORM_MODAL_BLOCK_ID]: {
+                [HITL_FREEFORM_MODAL_ACTION_ID]: { value: "approved with context" },
+              },
+            },
+          },
+        },
+      }),
+    );
+
+    expect(send).toHaveBeenCalledTimes(1);
+    const [payload, options] = send.mock.calls[0]!;
+    expect(payload).toEqual({
+      inputResponses: [{ requestId: "call_abc123", text: "approved with context" }],
+    });
+    expect(options).toMatchObject({
+      auth: {
+        attributes: {
+          author_type: "user",
+          channel_id: "C01",
+          team_id: "T01",
+          thread_ts: "1700000000.000001",
+          user_id: "U_SUBMITTER",
+          user_name: "grace",
+        },
+        authenticator: "slack-webhook",
+        issuer: "slack:T01",
+        principalId: "slack:T01:U_SUBMITTER",
+        principalType: "user",
+      },
+      continuationToken: "C01:1700000000.000001",
+      state: {
+        channelId: "C01",
+        teamId: "T01",
+        threadTs: "1700000000.000001",
+        triggeringUserId: "U_SUBMITTER",
+      },
+    });
+  });
+
+  it("rejects freeform modal answers from a different Slack user", async () => {
+    const channel = slackChannel({ credentials: { botToken: "xoxb-test" } });
+
+    const { send } = await firePost(
+      channel,
+      buildSignedInteractionRequest(
+        buildFreeformSubmissionPayload({
+          actorUserId: "U_INTRUDER",
+          responderUserId: "U_OWNER",
+          text: "delete it",
+        }),
+      ),
+    );
+
+    expect(send).not.toHaveBeenCalled();
+    expect(fetchMock).toHaveBeenCalledTimes(1);
+    const [url, init] = fetchMock.mock.calls[0]!;
+    expect(String(url)).toBe("https://slack.com/api/chat.postEphemeral");
+    expect(parseSlackRequestBody(init as RequestInit)).toMatchObject({
+      channel: "C01",
+      thread_ts: "1700000000.000001",
+      user: "U_INTRUDER",
+    });
   });
 });
